refactor: move decision manager code to a new file
